@@ -2,10 +2,10 @@ using System.Reflection;
 using System.Runtime.InteropServices;
 
 [assembly: AssemblyTitle(".NET Unified Platform")]
-[assembly: AssemblyVersion("3.0.0.0")]
+[assembly: AssemblyVersion("20.0.0.0")]
 [assembly: AssemblyCopyright("Copyright (c) 2024, Eben Roux")]
 [assembly: AssemblyProduct("Shuttle.OAuth")]
 [assembly: AssemblyCompany("Eben Roux")]
 [assembly: AssemblyConfiguration("Release")]
-[assembly: AssemblyInformationalVersion("3.0.0")]
+[assembly: AssemblyInformationalVersion("20.0.0")]
 [assembly: ComVisible(false)]
